backend/requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
motor==3.6.0
beanie==1.27.0
pydantic==2.7.4
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0
aiosmtplib==3.0.2
email-validator==2.2.0
python-dotenv==1.0.1
fpdf2==2.8.1
